Added Hide and updated HUD (on old code)
Had to roll back @tmoss343 changes, couldn't get it to work (check it
out maybe I'm missing something) I added a "Hide" mechanic and updated
the HUD. now the prompt changes.
