+++ .gitignore
@@ -23,3 +23,4 @@ Art/Concept_Art/shadow_creatures/Iconr
 Art/reference_art/Iconr
 Art/reference_art/City references /Iconr
 TechArt/ExternalTools/.coverage
+SoteriaGame/Library
