+++ SoteriaGame/Assembly-CSharp.csproj
@@ -46,6 +46,7 @@
 	</Reference>
   </ItemGroup>
   <ItemGroup>
+     <Compile Include="Assets\HidePlayer.cs" />
      <Compile Include="Assets\InstructionPrompt.cs" />
      <Compile Include="Assets\Scripts\CameraFollow.cs" />
      <Compile Include="Assets\Scripts\CollisionEvent.cs" />
