+++ SoteriaGame/Assets/EnemyMaterial.mat
@@ -19,6 +19,13 @@ Material:
         second:
           m_Texture: {fileID: 2800000, guid: f8a40ccd852537b4b875b66c0a211bb9, type: 3}
           m_Scale: {x: 1, y: 1}
+          m_Offset: {x: 1, y: 0}
+      data:
+        first:
+          name: _BumpMap
+        second:
+          m_Texture: {fileID: 0}
+          m_Scale: {x: 1, y: 1}
           m_Offset: {x: 0, y: 0}
       data:
         first:
@@ -27,7 +34,18 @@ Material:
           m_Texture: {fileID: 0}
           m_Scale: {x: 1, y: 1}
           m_Offset: {x: 0, y: 0}
-    m_Floats: {}
+      data:
+        first:
+          name: _ParallaxMap
+        second:
+          m_Texture: {fileID: 0}
+          m_Scale: {x: 1, y: 1}
+          m_Offset: {x: 0, y: 0}
+    m_Floats:
+      data:
+        first:
+          name: _Parallax
+        second: .0199999996
     m_Colors:
       data:
         first:
