+++ SoteriaGame/Assets/HidePlayer.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f1a785edea3a3804cad611161f24e0f5
+guid: 903d52360f412264695c52014f4621ff
 MonoImporter:
   serializedVersion: 2
   defaultReferences: []
