+++ SoteriaGame/Assets/SoteriaStatue.prefab
@@ -13,7 +13,6 @@ GameObject:
   - 23: {fileID: 2381762}
   - 108: {fileID: 10811016}
   - 114: {fileID: 11481428}
-  - 54: {fileID: 5440748}
   m_Layer: 0
   m_Name: SoteriaStatue
   m_TagString: Soteria Statue
@@ -60,21 +59,6 @@ MeshFilter:
   m_PrefabInternal: {fileID: 100100000}
   m_GameObject: {fileID: 181762}
   m_Mesh: {fileID: 10202, guid: 0000000000000000e000000000000000, type: 0}
---- !u!54 &5440748
-Rigidbody:
-  m_ObjectHideFlags: 1
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 181762}
-  serializedVersion: 2
-  m_Mass: 1
-  m_Drag: 0
-  m_AngularDrag: .0500000007
-  m_UseGravity: 0
-  m_IsKinematic: 0
-  m_Interpolate: 0
-  m_Constraints: 0
-  m_CollisionDetection: 0
 --- !u!65 &6581762
 BoxCollider:
   m_ObjectHideFlags: 1
