+++ SoteriaGame/SoteriaGame.userprefs
@@ -1,15 +1,6 @@
 ﻿<Properties>
-  <MonoDevelop.Ide.Workspace ActiveConfiguration="Debug" PreferredExecutionTarget="MonoDevelop.Default" />
-  <MonoDevelop.Ide.Workbench ActiveDocument="Assets\Scripts\Controllers\PCController.cs">
-    <Files>
-      <File FileName="Assets\Scripts\GameDirector.cs" Line="51" Column="5" />
-      <File FileName="Assets\Scripts\Controllers\PCController.cs" Line="85" Column="3" />
-      <File FileName="Assets\Scripts\Controllers\BasicEnemyController.cs" Line="1" Column="1" />
-      <File FileName="Assets\Scripts\Movement\BasicAggro.cs" Line="31" Column="50" />
-      <File FileName="Assets\Scripts\CollisionEvent.cs" Line="17" Column="6" />
-      <File FileName="Assets\Scripts\Controllers\Clickable.cs" Line="24" Column="35" />
-    </Files>
-  </MonoDevelop.Ide.Workbench>
+  <MonoDevelop.Ide.Workspace ActiveConfiguration="Debug" />
+  <MonoDevelop.Ide.Workbench />
   <MonoDevelop.Ide.DebuggingService.Breakpoints>
     <BreakpointStore />
   </MonoDevelop.Ide.DebuggingService.Breakpoints>
